Change ZipArtifacts to ZipAppInstaller
